+++ taxtastic/subcommands/update_taxids.py
@@ -1,8 +1,10 @@
-#!/usr/bin/env python
+"""Update obsolete tax_ids
+
+Use in preparation for ``taxit taxtable``. Takes sequence info file as
+passed to ``taxit create --seq-info``
+
 """
-Update obsolete tax_ids in preparation for use in taxtable. Takes sequence info
-file as passed to `taxit create --seq-info`
-"""
+
 # This file is part of taxtastic.
 #
 #    taxtastic is free software: you can redistribute it and/or modify
@@ -17,6 +19,7 @@ file as passed to `taxit create --seq-info`
 #
 #    You should have received a copy of the GNU General Public License
 #    along with taxtastic.  If not, see <http://www.gnu.org/licenses/>.
+
 import argparse
 import csv
 import logging
@@ -30,20 +33,28 @@ from taxtastic import ncbi
 
 log = logging.getLogger(__name__)
 
+
 def build_parser(parser):
-    parser.add_argument('infile', help="""Input CSV file to process, minimally
-            containing the fields 'seqname' and 'tax_id'. Rows with missing
-            tax_ids are left unchanged.""", type=argparse.FileType('r'))
-    parser.add_argument('-d', '--database-file', help="""Path to the taxonomy
-            database""", required=True)
-    parser.add_argument('-o', '--out-file', help="""Output file to write
-            updates [default: stdout]""", type=argparse.FileType('w'),
-            default=sys.stdout)
-    parser.add_argument('--unknowns', help="""csv file with single column 'seqname'
-            identifying records with unknown taxids (only if --unknown-action=remove)""", type=argparse.FileType('w'))
-    parser.add_argument('-u', '--unknown-action', help="""Action to take on
-            encountering an unknown tax_id [default: %(default)s]""",
-            choices=('halt', 'remove'), default='halt')
+    parser.add_argument(
+        'infile', type=argparse.FileType('r'),
+        help="""Input CSV file to process, minimally containing the
+        fields 'seqname' and 'tax_id'. Rows with missing tax_ids are
+        left unchanged.""")
+    parser.add_argument(
+        '-d', '--database-file', required=True,
+        help="""Path to the taxonomy database""")
+    parser.add_argument(
+        '-o', '--out-file', type=argparse.FileType('w'), default=sys.stdout,
+        help="""Output file to write updates [default: stdout]""")
+    parser.add_argument(
+        '--unknowns', type=argparse.FileType('w'),
+        help="""csv file with single column 'seqname' identifying
+        records with unknown taxids (only if
+        --unknown-action=remove)""")
+    parser.add_argument(
+        '-u', '--unknown-action', choices=('halt', 'remove'), default='halt',
+        help="""Action to take on encountering an unknown tax_id
+        [default: %(default)s]""")
 
 
 def load_csv(fp):
@@ -57,6 +68,7 @@ def load_csv(fp):
     dialect = csv.Sniffer().sniff(sample)
     reader = csv.DictReader(fp, dialect=dialect)
     return (reader.fieldnames, dialect, reader)
+
 
 def taxid_updater(taxonomy, action='halt'):
     """
@@ -79,7 +91,7 @@ def taxid_updater(taxonomy, action='halt'):
         if new_tax_id and new_tax_id != current_tax_id:
             row['tax_id'] = new_tax_id
             log.warn('Replacing %s with %s [%s]', current_tax_id, new_tax_id,
-                    row['seqname'])
+                     row['seqname'])
         elif action == 'halt':
             raise KeyError("Unknown taxon {0}".format(current_tax_id))
         elif action == 'remove':
@@ -115,7 +127,8 @@ def action(args):
 
     if args.unknowns:
         # unknown taxids are set to empty string in taxid_updater
-        csv.writer(args.unknowns).writerows([r['seqname']] for r in updated if not r['tax_id'])
+        csv.writer(args.unknowns).writerows(
+            [r['seqname']] for r in updated if not r['tax_id'])
 
     with args.out_file as fp:
         writer = csv.DictWriter(fp, headers, dialect)